Add parser to get the issuer of satellite client cert
* Add parser to get issuer of "/etc/rhsm/ca/katello-default-ca.pem".
* Refactor CertificateChain to reuse the CertificateInfo since chain is just
  a list of cert.
