--- docs/shared_parsers_catalog/certificate_chain.rst
@@ -1,3 +1,0 @@
-.. automodule:: insights.parsers.certificate_chain
-   :members:
-   :show-inheritance:
