+++ docs/shared_parsers_catalog/ssl_certificate.rst
@@ -1,0 +1,3 @@
+.. automodule:: insights.parsers.ssl_certificate
+   :members:
+   :show-inheritance:
